+++ kbase-extension/static/kbase/js/common/monoBus.js
@@ -14,7 +14,7 @@
  *
  *
  */
-define(['uuid', 'bluebird', './lang', './unodep'], (Uuid, Promise, lang, utils) => {
+define(['uuid', 'bluebird', './lang'], (Uuid, Promise, lang) => {
     'use strict';
 
     function factory(cfg) {
@@ -349,7 +349,7 @@ define(['uuid', 'bluebird', './lang', './unodep'], (Uuid, Promise, lang, utils) 
 
             const existingMessage = channel.persistentMessages[key];
             if (existingMessage) {
-                if (utils.isEqual(existingMessage.message, message)) {
+                if (lang.isEqual(existingMessage.message, message)) {
                     return;
                 }
             }
